perf(viewer): POSITIVE_SCALAR semantic → Float16 on GPU
Phase-4 C-ts-3: activate Float16 GPU dtype for POSITIVE_SCALAR
attributes — radii (Points), widths + segmentLength (Lines),
amplitudes (GSplats). Disk format unchanged; narrowing happens in
the packer at upload time.

Map flip: `DEFAULT_GPU_DTYPE_BY_SEMANTIC.positive_scalar` flipped
from `'float32'` to `'float16'`. Single-line change — the C-ts-1
foundation routes positive-scalar specs through the existing
narrowing path automatically.

Why no shader changes: Float16 is a STORAGE dtype. The shader still
reads `float`; the GPU widens to single precision on attribute
fetch. Values pass through unchanged within Float16 ULP
(~5e-4 relative). Existing per-material `radiusScale` /
`sharpnessScale` uniforms (default 1.0) continue to apply.

BOUNDED_SCALAR (sharpness, clipped) stays at Float32 in this
commit — Uint8-norm narrowing requires threading the encoder's
bounds metadata (V_max) to the shader's existing scale uniforms.
Folded into a future C-ts-3b sub-commit (or skip if perf gain is
marginal vs complexity).

Bandwidth impact: 2 bytes/inst saved per POSITIVE_SCALAR attribute.
- Points (aRadius): 2 bytes/inst saved.
- Lines (aStartWidth + aEndWidth + aSegmentLength): 6 bytes/inst saved.
- GSplats (aAmplitude): 2 bytes/inst saved.

Combined with C-ts-2 COLOR narrowing:
- Points: 8 bytes/inst saved (color + radius).
- Lines: 18 bytes/inst saved (colors + widths + segmentLength).
- GSplats: 8 bytes/inst saved (color + amplitude).

On the 10M-line synthetic scenario: ~180 MB GPU VRAM saved
cumulative across C-ts-2 + C-ts-3.

Gates:
- pnpm typecheck: clean
- pnpm test --run: 4740 passed
- tsl-shader-parity: 14/14 pass (Float16 ULP on scene-unit
  magnitudes ~5e-4 — well below mean-abs-diff threshold)
- tsl-codegen-snapshot: zero diff
- E2E smoke (basic-rendering + geometry-types): 9/9 pass

+++ packages/luxar-viewer/src/rendering/interleaved-attributes.ts
@@ -176,7 +176,18 @@ const DEFAULT_GPU_DTYPE_BY_SEMANTIC: Record<SemanticType, GpuDtype> = {
   // ~5e-4 absolute error on SDR colours in [0,1] — well below the
   // parity-spec mean-abs-diff threshold on 0-255 scale.
   color: 'float16',
-  positive_scalar: 'float32',
+  // C-ts-3: POSITIVE_SCALAR (radii, widths, amplitudes) narrowed to
+  // Float16 on GPU. These hold scene-unit magnitudes; Float16 covers
+  // up to ±65504 with ~5e-4 relative precision — overkill for scene
+  // distances and far more headroom than needed for typical scenes.
+  // No shader change: TSL/GLSL still read `float`. GPU widens at
+  // attribute fetch.
+  positive_scalar: 'float16',
+  // C-ts-3b (future): BOUNDED_SCALAR (sharpness, clipped flags)
+  // will narrow to Uint8-norm. Requires threading the encoder's
+  // bounds metadata (V_max) to the shader's `sharpnessScale` /
+  // `radiusScale` uniforms — non-trivial because bounds vary per
+  // dataset. Stays Float32 for now.
   bounded_scalar: 'float32',
   cholesky: 'float32',
   // INDEX attributes don't flow through this packer (they're handled
